requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
fastapi
uvicorn
pydantic
requests
python-dotenv
Pillow
mutagen
